LPS-54322 Move EditPasswordPolicyAssignmentsAction.java, no logic changes
