+++ portal-impl/src/com/liferay/portlet/passwordpoliciesadmin/PasswordPoliciesAdminPortlet.java
@@ -17,18 +17,26 @@ package com.liferay.portlet.passwordpoliciesadmin;
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCPortlet;
 
 import com.liferay.portal.DuplicatePasswordPolicyException;
+import com.liferay.portal.kernel.servlet.SessionErrors;
+import com.liferay.portal.kernel.util.Constants;
+import com.liferay.portal.kernel.util.ParamUtil;
+import com.liferay.portal.kernel.util.StringUtil;
+import com.liferay.portal.kernel.util.Validator;
+import com.liferay.portal.model.PasswordPolicy;
 import com.liferay.portal.NoSuchPasswordPolicyException;
 import com.liferay.portal.PasswordPolicyNameException;
 import com.liferay.portal.RequiredPasswordPolicyException;
-import com.liferay.portal.kernel.util.ParamUtil;
-import com.liferay.portal.model.PasswordPolicy;
 import com.liferay.portal.security.auth.PrincipalException;
+import com.liferay.portal.service.OrganizationServiceUtil;
 import com.liferay.portal.service.PasswordPolicyServiceUtil;
 import com.liferay.portal.service.ServiceContext;
 import com.liferay.portal.service.ServiceContextFactory;
+import com.liferay.portal.service.UserServiceUtil;
+import com.liferay.portal.struts.PortletAction;
 
 import javax.portlet.ActionRequest;
 import javax.portlet.ActionResponse;
+import javax.portlet.PortletConfig;
 import javax.portlet.RenderRequest;
 import javax.portlet.RenderResponse;
 
@@ -36,6 +44,110 @@ import javax.portlet.RenderResponse;
  * @author Drew Brokke
  */
 public class PasswordPoliciesAdminPortlet extends MVCPortlet {
+
+	@Override
+	public void processAction(
+			ActionMapping actionMapping, ActionForm actionForm,
+			PortletConfig portletConfig, ActionRequest actionRequest,
+			ActionResponse actionResponse)
+		throws Exception {
+
+		String cmd = ParamUtil.getString(actionRequest, Constants.CMD);
+
+		try {
+			if (cmd.equals("password_policy_organizations")) {
+				updatePasswordPolicyOrganizations(actionRequest);
+			}
+			else if (cmd.equals("password_policy_users")) {
+				updatePasswordPolicyUsers(actionRequest);
+			}
+
+			if (Validator.isNotNull(cmd)) {
+				String redirect = ParamUtil.getString(
+					actionRequest, "assignmentsRedirect");
+
+				sendRedirect(actionRequest, actionResponse, redirect);
+			}
+		}
+		catch (Exception e) {
+			if (e instanceof NoSuchPasswordPolicyException ||
+				e instanceof PrincipalException) {
+
+				SessionErrors.add(actionRequest, e.getClass());
+
+				setForward(
+					actionRequest, "portlet.password_policies_admin.error");
+			}
+			else {
+				throw e;
+			}
+		}
+	}
+
+	@Override
+	public ActionForward render(
+			ActionMapping actionMapping, ActionForm actionForm,
+			PortletConfig portletConfig, RenderRequest renderRequest,
+			RenderResponse renderResponse)
+		throws Exception {
+
+		try {
+			ActionUtil.getPasswordPolicy(renderRequest);
+		}
+		catch (Exception e) {
+			if (e instanceof NoSuchPasswordPolicyException ||
+				e instanceof PrincipalException) {
+
+				SessionErrors.add(renderRequest, e.getClass());
+
+				return actionMapping.findForward(
+					"portlet.password_policies_admin.error");
+			}
+			else {
+				throw e;
+			}
+		}
+
+		return actionMapping.findForward(
+			getForward(
+				renderRequest,
+				"portlet.password_policies_admin." +
+					"edit_password_policy_assignments"));
+	}
+
+	protected void updatePasswordPolicyOrganizations(
+			ActionRequest actionRequest)
+		throws Exception {
+
+		long passwordPolicyId = ParamUtil.getLong(
+			actionRequest, "passwordPolicyId");
+
+		long[] addOrganizationIds = StringUtil.split(
+			ParamUtil.getString(actionRequest, "addOrganizationIds"), 0L);
+		long[] removeOrganizationIds = StringUtil.split(
+			ParamUtil.getString(actionRequest, "removeOrganizationIds"), 0L);
+
+		OrganizationServiceUtil.addPasswordPolicyOrganizations(
+			passwordPolicyId, addOrganizationIds);
+		OrganizationServiceUtil.unsetPasswordPolicyOrganizations(
+			passwordPolicyId, removeOrganizationIds);
+	}
+
+	protected void updatePasswordPolicyUsers(ActionRequest actionRequest)
+		throws Exception {
+
+		long passwordPolicyId = ParamUtil.getLong(
+			actionRequest, "passwordPolicyId");
+
+		long[] addUserIds = StringUtil.split(
+			ParamUtil.getString(actionRequest, "addUserIds"), 0L);
+		long[] removeUserIds = StringUtil.split(
+			ParamUtil.getString(actionRequest, "removeUserIds"), 0L);
+
+		UserServiceUtil.addPasswordPolicyUsers(passwordPolicyId, addUserIds);
+		UserServiceUtil.unsetPasswordPolicyUsers(
+			passwordPolicyId, removeUserIds);
+	}
 
 	public void deletePasswordPolicy(
 		ActionRequest actionRequest, ActionResponse actionResponse)
