--- portal-impl/src/com/liferay/portlet/passwordpoliciesadmin/action/EditPasswordPolicyAssignmentsAction.java
@@ -1,147 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.portlet.passwordpoliciesadmin.action;
-
-import com.liferay.portal.NoSuchPasswordPolicyException;
-import com.liferay.portal.kernel.servlet.SessionErrors;
-import com.liferay.portal.kernel.util.Constants;
-import com.liferay.portal.kernel.util.ParamUtil;
-import com.liferay.portal.kernel.util.StringUtil;
-import com.liferay.portal.kernel.util.Validator;
-import com.liferay.portal.security.auth.PrincipalException;
-import com.liferay.portal.service.OrganizationServiceUtil;
-import com.liferay.portal.service.UserServiceUtil;
-import com.liferay.portal.struts.PortletAction;
-
-import javax.portlet.ActionRequest;
-import javax.portlet.ActionResponse;
-import javax.portlet.PortletConfig;
-import javax.portlet.RenderRequest;
-import javax.portlet.RenderResponse;
-
-import org.apache.struts.action.ActionForm;
-import org.apache.struts.action.ActionForward;
-import org.apache.struts.action.ActionMapping;
-
-/**
- * @author Scott Lee
- */
-public class EditPasswordPolicyAssignmentsAction extends PortletAction {
-
-	@Override
-	public void processAction(
-			ActionMapping actionMapping, ActionForm actionForm,
-			PortletConfig portletConfig, ActionRequest actionRequest,
-			ActionResponse actionResponse)
-		throws Exception {
-
-		String cmd = ParamUtil.getString(actionRequest, Constants.CMD);
-
-		try {
-			if (cmd.equals("password_policy_organizations")) {
-				updatePasswordPolicyOrganizations(actionRequest);
-			}
-			else if (cmd.equals("password_policy_users")) {
-				updatePasswordPolicyUsers(actionRequest);
-			}
-
-			if (Validator.isNotNull(cmd)) {
-				String redirect = ParamUtil.getString(
-					actionRequest, "assignmentsRedirect");
-
-				sendRedirect(actionRequest, actionResponse, redirect);
-			}
-		}
-		catch (Exception e) {
-			if (e instanceof NoSuchPasswordPolicyException ||
-				e instanceof PrincipalException) {
-
-				SessionErrors.add(actionRequest, e.getClass());
-
-				setForward(
-					actionRequest, "portlet.password_policies_admin.error");
-			}
-			else {
-				throw e;
-			}
-		}
-	}
-
-	@Override
-	public ActionForward render(
-			ActionMapping actionMapping, ActionForm actionForm,
-			PortletConfig portletConfig, RenderRequest renderRequest,
-			RenderResponse renderResponse)
-		throws Exception {
-
-		try {
-			ActionUtil.getPasswordPolicy(renderRequest);
-		}
-		catch (Exception e) {
-			if (e instanceof NoSuchPasswordPolicyException ||
-				e instanceof PrincipalException) {
-
-				SessionErrors.add(renderRequest, e.getClass());
-
-				return actionMapping.findForward(
-					"portlet.password_policies_admin.error");
-			}
-			else {
-				throw e;
-			}
-		}
-
-		return actionMapping.findForward(
-			getForward(
-				renderRequest,
-				"portlet.password_policies_admin." +
-					"edit_password_policy_assignments"));
-	}
-
-	protected void updatePasswordPolicyOrganizations(
-			ActionRequest actionRequest)
-		throws Exception {
-
-		long passwordPolicyId = ParamUtil.getLong(
-			actionRequest, "passwordPolicyId");
-
-		long[] addOrganizationIds = StringUtil.split(
-			ParamUtil.getString(actionRequest, "addOrganizationIds"), 0L);
-		long[] removeOrganizationIds = StringUtil.split(
-			ParamUtil.getString(actionRequest, "removeOrganizationIds"), 0L);
-
-		OrganizationServiceUtil.addPasswordPolicyOrganizations(
-			passwordPolicyId, addOrganizationIds);
-		OrganizationServiceUtil.unsetPasswordPolicyOrganizations(
-			passwordPolicyId, removeOrganizationIds);
-	}
-
-	protected void updatePasswordPolicyUsers(ActionRequest actionRequest)
-		throws Exception {
-
-		long passwordPolicyId = ParamUtil.getLong(
-			actionRequest, "passwordPolicyId");
-
-		long[] addUserIds = StringUtil.split(
-			ParamUtil.getString(actionRequest, "addUserIds"), 0L);
-		long[] removeUserIds = StringUtil.split(
-			ParamUtil.getString(actionRequest, "removeUserIds"), 0L);
-
-		UserServiceUtil.addPasswordPolicyUsers(passwordPolicyId, addUserIds);
-		UserServiceUtil.unsetPasswordPolicyUsers(
-			passwordPolicyId, removeUserIds);
-	}
-
-}
